requirements.txt
streamlit
openai
serpapi
python-dotenv
google-search-results
